TODO

+++ mf-deliver-client/src/main/java/com/mfexpress/rent/deliver/dto/data/serve/ServeDTO.java
@@ -43,7 +43,7 @@ public class ServeDTO {
 
     private BigDecimal rent;
 
-    private Double deposit;
+    private BigDecimal deposit;
 
     private Date createTime;
